Hide comparator and sortSpecification inside QuerySort
