+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/extensions/ClientExtensions.kt
@@ -1,6 +1,5 @@
 package io.getstream.chat.android.livedata.extensions
 
-import io.getstream.chat.android.client.api.models.QuerySort
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.errors.ChatNetworkError
 import io.getstream.chat.android.client.models.Channel
@@ -11,9 +10,6 @@ import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.PerformanceUtils
 import io.getstream.chat.android.livedata.request.AnyChannelPaginationRequest
-import kotlin.reflect.KProperty1
-
-private const val EQUAL_ON_COMPARISON = 0
 
 internal fun Message.users(): List<User> = latestReactions.mapNotNull(Reaction::user) + user
 
@@ -142,32 +138,5 @@ internal fun Collection<Channel>.applyPagination(pagination: AnyChannelPaginatio
         val comparator = PerformanceUtils.task("Get comparator") { pagination.sort.comparator }
         PerformanceUtils.task("Sorting") { sortedWith(comparator) }
     }.drop(pagination.channelOffset).take(pagination.channelLimit).toList()
-
-internal val QuerySort<Channel>.comparator: Comparator<in Channel>
-    get() = CompositeComparator(this.sortSpecifications.mapNotNull { it.comparator as? Comparator<Channel> })
-
-internal val QuerySort.SortSpecification<Channel>.comparator: Comparator<in Channel>?
-    get() = (this.sortAttribute as? QuerySort.SortAttribute.FieldSortAttribute<Channel>)?.field?.comparator(this.sortDirection)
-
-internal fun KProperty1<Channel, Comparable<*>?>.comparator(sortDirection: QuerySort.SortDirection): Comparator<Channel>? =
-    this.let { compareProperty ->
-        Comparator { c0, c1 ->
-            (compareProperty.getter.call(c0) as? Comparable<Any>)?.let { a ->
-                (compareProperty.getter.call(c1) as? Comparable<Any>)?.let { b ->
-                    a.compareTo(b) * sortDirection.value
-                }
-            } ?: EQUAL_ON_COMPARISON
-        }
-    }
-
-internal class CompositeComparator<T>(private val comparators: List<Comparator<T>>) : Comparator<T> {
-    override fun compare(o1: T, o2: T): Int =
-        comparators.fold(EQUAL_ON_COMPARISON) { currentComparisonValue, comparator ->
-            when (currentComparisonValue) {
-                EQUAL_ON_COMPARISON -> comparator.compare(o1, o2)
-                else -> currentComparisonValue
-            }
-        }
-}
 
 internal fun String?.isImageMimetype() = this?.contains("image") ?: false
